only on windows

+++ tmp_del.sh
@@ -1,2 +1,4 @@
-cd $TMP
-rm -rf *
+if [[ "$OSTYPE" == "msys" ]]; then
+	cd $TMP
+	rm -rf *
+fi
